fixed issue that occur when mkfs calculate rootfs size (batch updates)

+++ tools/update_prebuilt.sh
@@ -6,10 +6,12 @@ set -eu
 	exit 1
 }
 
-USERDATA_SIZE=104857600
-echo "Generating empty userdata.img (size:${USERDATA_SIZE})"
-TMPDIR=`mktemp -d`
-${PWD}/tools/make_ext4fs -s -l ${USERDATA_SIZE} -a root -L userdata $1/userdata.img ${TMPDIR}
-rm -rf ${TMPDIR}
+if [ ! -f $1/userdata.img ]; then
+	USERDATA_SIZE=104857600
+	echo "Generating empty userdata.img (size:${USERDATA_SIZE})"
+	TMPDIR=`mktemp -d`
+	${PWD}/tools/make_ext4fs -s -l ${USERDATA_SIZE} -a root -L userdata $1/userdata.img ${TMPDIR}
+	rm -rf ${TMPDIR}
+fi
 
 exit $?
